requirements/base.txt
django==4.0.6
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
requests
django-ckeditor==6.7.1
django-resized==1.0.2
django-import-export==4.1.1
django-import-export[xlsx]
python-decouple==3.8